fix: address code review issues for mDoc support
- Add null checks for credentialSubject in MdocCredentialBuilder
- Guard CBORGenerator cast with instanceof check in Tagged and IntegerMap
- Throw MdocException instead of returning null for missing proof key algorithm
- Change getCoseAlgorithmIdentifier return type from int to long
- Remove redundant catch-and-rethrow MdocException block

+++ core/src/main/java/org/keycloak/mdoc/MdocAlgorithm.java
@@ -53,8 +53,8 @@ public enum MdocAlgorithm {
         return joseAlgorithm;
     }
 
-    public int getCoseAlgorithmIdentifier() {
-        return (int) coseAlgorithmIdentifier.getValue();
+    public long getCoseAlgorithmIdentifier() {
+        return coseAlgorithmIdentifier.getValue();
     }
 
     public COSEAlgorithmIdentifier toCoseAlgorithmIdentifier() {
